requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pymongo==4.6.1
redis==5.0.1
google-cloud-speech==2.24.0
google-cloud-texttospeech==2.16.0
google-generativeai==0.3.2
httpx==0.26.0
aiofiles==23.2.1
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
structlog==24.1.0
gunicorn>=21.2.0
uvloop>=0.19.0
